vibraphone to class component (i'm done with func)
